\name{TmpLasso-package}
\alias{TmpLasso-package}
\alias{TmpLasso}
\docType{package}
\title{\packageTitle{TmpLasso}}
\description{\packageDescription{TmpLasso}}
\section{Package Content}{\packageIndices{TmpLasso}}
\author{\packageAuthor{TmpLasso}}
\section{Maintainer}{\packageMaintainer{TmpLasso}}
\keyword{package}
